Silviu/cos 4165 notes and relationship in the automated test (#8342)
* [4165] timeline and refactoring

* [4165] timeline and refactoring

+++ packages/apps/frontera/tests/test.spec.ts
@@ -4,6 +4,7 @@ import { CustomersPage } from './pages/customersPage';
 import { AddressBookPage } from './pages/addressBookPage';
 import { OrganizationAccountPage } from './pages/organization/organizationAccountPage';
 import { OrganizationSideNavPage } from './pages/organization/organizationSideNavPage';
+import { OrganizationTimelinePage } from './pages/organization/organizationTimelinePage';
 
 test.setTimeout(180000);
 
@@ -41,6 +42,7 @@ test('create and delete contracts', async ({ page }) => {
   const addressBookPage = new AddressBookPage(page);
   const organizationAccountPage = new OrganizationAccountPage(page);
   const organizationSideNavPage = new OrganizationSideNavPage(page);
+  const organizationTimelinePage = new OrganizationTimelinePage(page);
 
   // Login
   await loginPage.login();
@@ -50,18 +52,28 @@ test('create and delete contracts', async ({ page }) => {
   // Add organization and check new entry
   await addressBookPage.addOrganization();
 
-  // Add contract to organization and check new entry
   await new Promise((resolve) => setTimeout(resolve, 1500));
-  // await page.reload();
   await addressBookPage.goToOrganization();
   await organizationSideNavPage.goToAccount();
+  await organizationAccountPage.updateOrgToCustomer();
+  await organizationAccountPage.addNoteToOrg();
+  await organizationTimelinePage.ensureEmailPermissionPromptIsRedirecting();
+  await page.goBack();
+  await organizationTimelinePage.ensureLogEntryCanBeAdded();
+  await organizationTimelinePage.ensureReminderCanBeAdded();
+
+  // Add the first contract to organization and check new entry
   await organizationAccountPage.addContractEmpty();
   await organizationAccountPage.addBillingAddress(0);
   await organizationAccountPage.checkContractsCount(1);
   await organizationAccountPage.addSLIsToContract(0);
   await organizationAccountPage.checkSLIsInAccountPanel();
+
+  // Add the second first contract to organization
   await organizationAccountPage.addContractNonEmpty();
   await organizationAccountPage.checkContractsCount(2);
+
+  // Delete a contract
   await organizationAccountPage.deleteContract(1);
   await organizationAccountPage.checkContractsCount(1);
 });
